requirements.txt
Django==4.0.4
python
whitenoise==6.1.0
django-bootstrap-datepicker-plus==4.0.0
django-import_export==2.8.0
dill==0.3.4
pandas==1.4.2
requests==2.27.1
python-docx==0.8.11
psycopg2==2.9.3
pypdf4==1.27.0
pywin32
psycopg2-binary
easygui==0.98.3
numpy==1.22.3
nltk==3.7
spacy==3.3.0
pikepdf==5.1.3
pillow==9.1.1
fsspec